Change article-header style, hide tag/category seperators if they're not present
The latter point requires some horrible hacks.

As I can't access tags from Pollen, detecting whether tags are empty
or not has to be done in Frog templates. This is done by having
◊article-header pass Frog template code as string output in what is
supposed to be HTML:

```racket
;; ◊article-header in pollen.rkt
(->html
 `(header ([class ,class])
          ; ...
          (p ([class "date-and-category"])
             (span ,date)
             ,(string-append "@(when " category " \", \")")
             ,category
             ,(string-append "@(when " tags " \" :: \")")
             ,tags))))
```

`(string-append "@(when " category "\", \")")` gets Frog code from the
function input as `category` and spits out Frog code like `@(when
@|category-from-tags| ",")` as a string.

This also relies on `category` and `tags` being set to #f if they turn
out to be empty. The relevant logic is in `tags.rkt`: use `and~>` so
that a #f in the "pipe" will short-circuit, then return #f in the
"pipe" when the intermediate list turns out to be empty.

```racket
;; tags->comma-html
(and~> ...
       ;; The "#f if empty" "pipe" function
       ((λ (lst) (if (empty? lst) #f lst)) _)
       ...)
```

Then index-template & post-template have to use and~> when processing
filtered-tags and category-from-tags so the #f from tags->comma-html
also short-circuits them, setting them to #f to let the @when know
whether to return the string or not.

All just to hide the seperators if there's no tag/category.