Use body font in figcaption
It's not different enough from body to be highlighted with a different
font.

Previously I had to do this because with the <p> wrapping the image,
it was hard to distinguish the figcaption from the rest of the
article. But that is no longer an issue after we stopped Pollen from
adding the <p> in <figure> tags.