link directly to the RSS feed
previously I thought it'd be a good idea if I show a page explaining
what RSS is instead of leading the viewer into a raw XML page. It...
is, but my explainer page wasn't well written.

I don't really know where to start explaining it, so I might as well
do what every other site does and assume that the viewer already knows
what RSS is before clicking it.