display tight lists tightly
tight lists are usually <li> without <p> in them. In Markdown or Org:

- item 1
- item 2

output:

<ul>
<li>item 1</li>
<li>item 2</li>
</ul>

It's better when the text is actually put into a <p> for consistency.
But I still want tight lists to appear as tight. We can remove the
top/bottom margins from the first <p> under an <li> to achieve that.