update tailwindcss to 3.1.6
Tailwind 3.1 integrates postcss-import so we don't have to install or
set it up ourselves.

postcss-cli is also not needed for the same reason, though this might
not have been a change from Tailwind 3.1.

postcss.config.js is still needed for tailwindcss/nesting.

Tailwind CLI still calls autoprefixer and does the right thing without
declaring it in postcss.config.js.