Switch to running TailwindCSS (PostCSS) through Hugo
Before this, editing classes would trigger both the Tailwind rebuild and
one browser reload from the HTML changing, but the browser would often
reload before the Tailwind rebuild was done, and so we end up having to
manually reload again before we see the definitions of new classes.

(This only shows up if a class previously not mentioned is now mentioned.)

When generating CSS through Hugo, the HTML change still triggers one
immediate browser reload, but then after the Tailwind rebuild the
browser is prompted to reload again. So it's not as clean as, say, a hot
reload in Astro, but at least it's more automatic.

A lot of the set up came from here:
https://github.com/bep/hugo-starter-tailwind-basic/blob/master/hugo.toml